requirements.txt
langchain
langchain-core
langchain-openai
openai
python-dotenv
pydantic